+++ integration_tests/tests/fixtures/vault_client.rs
@@ -1,5 +1,6 @@
 use std::mem::size_of;
 
+use crate::fixtures::TestError;
 use jito_account_traits::AccountDeserialize;
 use jito_restaking_core::{
     ncn_operator_ticket::NcnOperatorTicket, ncn_vault_slasher_ticket::NcnVaultSlasherTicket,
@@ -13,7 +14,7 @@ use jito_vault_core::{
     vault_operator_ticket::VaultOperatorTicket,
     vault_staker_withdrawal_ticket::VaultStakerWithdrawalTicket,
 };
-use jito_vault_sdk::{
+use jito_vault_sdk::sdk::{
     add_delegation, initialize_config, initialize_vault, initialize_vault_delegation_list,
 };
 use log::info;
@@ -36,8 +37,6 @@ use spl_associated_token_account::{
     get_associated_token_address, instruction::create_associated_token_account_idempotent,
 };
 use spl_token::{instruction::initialize_mint2, state::Mint};
-
-use crate::fixtures::TestError;
 
 pub struct VaultRoot {
     pub vault_pubkey: Pubkey,
@@ -598,7 +597,7 @@ impl VaultProgramClient {
         let blockhash = self.banks_client.get_latest_blockhash().await?;
 
         self._process_transaction(&Transaction::new_signed_with_payer(
-            &[jito_vault_sdk::initialize_vault_ncn_ticket(
+            &[jito_vault_sdk::sdk::initialize_vault_ncn_ticket(
                 &jito_vault_program::id(),
                 config,
                 vault,
@@ -627,7 +626,7 @@ impl VaultProgramClient {
     ) -> Result<(), TestError> {
         let blockhash = self.banks_client.get_latest_blockhash().await?;
         self._process_transaction(&Transaction::new_signed_with_payer(
-            &[jito_vault_sdk::initialize_vault_operator_ticket(
+            &[jito_vault_sdk::sdk::initialize_vault_operator_ticket(
                 &jito_vault_program::id(),
                 config,
                 vault,
@@ -704,7 +703,7 @@ impl VaultProgramClient {
         let vault = self.get_vault(vault_pubkey).await?;
 
         self._process_transaction(&Transaction::new_signed_with_payer(
-            &[jito_vault_sdk::update_vault(
+            &[jito_vault_sdk::sdk::update_vault(
                 &jito_vault_program::id(),
                 &Config::find_program_address(&jito_vault_program::id()).0,
                 vault_pubkey,
@@ -734,7 +733,7 @@ impl VaultProgramClient {
     ) -> Result<(), TestError> {
         let blockhash = self.banks_client.get_latest_blockhash().await?;
         self._process_transaction(&Transaction::new_signed_with_payer(
-            &[jito_vault_sdk::enqueue_withdraw(
+            &[jito_vault_sdk::sdk::enqueue_withdraw(
                 &jito_vault_program::id(),
                 config,
                 vault,
@@ -805,7 +804,7 @@ impl VaultProgramClient {
     ) -> Result<(), TestError> {
         let blockhash = self.banks_client.get_latest_blockhash().await?;
         self._process_transaction(&Transaction::new_signed_with_payer(
-            &[jito_vault_sdk::burn_withdrawal_ticket(
+            &[jito_vault_sdk::sdk::burn_withdrawal_ticket(
                 &jito_vault_program::id(),
                 config,
                 vault,
@@ -874,7 +873,7 @@ impl VaultProgramClient {
             signers.push(signer);
         }
         self._process_transaction(&Transaction::new_signed_with_payer(
-            &[jito_vault_sdk::mint_to(
+            &[jito_vault_sdk::sdk::mint_to(
                 &jito_vault_program::id(),
                 &Config::find_program_address(&jito_vault_program::id()).0,
                 vault,
@@ -907,7 +906,7 @@ impl VaultProgramClient {
     ) -> Result<(), TestError> {
         let blockhash = self.banks_client.get_latest_blockhash().await?;
         self._process_transaction(&Transaction::new_signed_with_payer(
-            &[jito_vault_sdk::initialize_vault_ncn_slasher_ticket(
+            &[jito_vault_sdk::sdk::initialize_vault_ncn_slasher_ticket(
                 &jito_vault_program::id(),
                 config,
                 vault,
@@ -939,7 +938,7 @@ impl VaultProgramClient {
         let blockhash = self.banks_client.get_latest_blockhash().await?;
         self._process_transaction(&Transaction::new_signed_with_payer(
             &[
-                jito_vault_sdk::initialize_vault_ncn_slasher_operator_ticket(
+                jito_vault_sdk::sdk::initialize_vault_ncn_slasher_operator_ticket(
                     &jito_vault_program::id(),
                     config,
                     vault,
@@ -981,7 +980,7 @@ impl VaultProgramClient {
     ) -> Result<(), TestError> {
         let blockhash = self.banks_client.get_latest_blockhash().await?;
         self._process_transaction(&Transaction::new_signed_with_payer(
-            &[jito_vault_sdk::slash(
+            &[jito_vault_sdk::sdk::slash(
                 &jito_vault_program::id(),
                 config,
                 vault,
