+++ vault_core/Cargo.toml
@@ -18,6 +18,7 @@ solana-program = { workspace = true }
 spl-associated-token-account = { workspace = true }
 spl-token = { workspace = true }
 thiserror = { workspace = true }
+jito-vault-sdk = { workspace = true }
 
 [dev-dependencies]
 assert_matches = { workspace = true }
