+++ vault_program/src/set_secondary_admin.rs
@@ -4,7 +4,7 @@ use jito_vault_core::{
     loader::{load_config, load_vault},
     vault::Vault,
 };
-use jito_vault_sdk::VaultAdminRole;
+use jito_vault_sdk::instruction::VaultAdminRole;
 use solana_program::{
     account_info::AccountInfo, entrypoint::ProgramResult, msg, program_error::ProgramError,
     pubkey::Pubkey,
